fix(types): boundary residue → arg-type 0 (arg-type PR K)
Clears the final 30 `mypy --enable-error-code arg-type core` errors, taking the
sweep to 0 (PR L flips the global default next). Cleared structurally — zero
`# type: ignore` overlays — by fixing real bugs, aligning protocols to their
impls, and typing heterogeneous boundaries honestly.

Real bugs surfaced + fixed:
- ps_ai_service: CORE-fallback unpacked `backend.search` (returns Result[list],
  not a (list, total) tuple) → would mis-unpack the list. Drop the unpack.
- tasks_intelligence: read non-existent key `prerequisites` from
  KnowledgePrerequisitesResult (real key: `required_knowledge`) → prereqs always
  empty / has_prerequisites always False. Plus None-guard graph_intel.
- query_processor: passed a dict as `user_context: str` to LLMService →
  `"\n".join` would crash; build a natural-language summary string instead.
- _spawn_orchestrator: reversed annotations on computed cross-edges
  (RelationshipName, str) — runtime already correct, types lied.

Protocol/impl alignment:
- HabitsSearchOperations.get_by_frequency: drop phantom `user_uid` param the
  concrete impl never had (was shifting `limit` onto it).
- GoalsSearchOperations.get_by_category: add the `user_uid` the concrete
  (BaseService) impl has; drop the masking `# type: ignore[call-arg]`.

Honest typing of boundaries:
- adaptive_lp_cross_domain: `_DomainSynergy` TypedDict for the synergy table.
- _alignment_intelligence_mixin: convert user-supplied level → AlignmentLevel
  (UNKNOWN sentinel for missing/invalid).
- lifepath_core_service: `dimension_scores` typed AlignmentDimensions (its only
  producer).
- search_request: `connected_direction` → Literal (Pydantic validates at edge).
- relationship_registry: USER_CONFIG uses real User/UserDTO (not None);
  bidirectional_relationships widened to the union it genuinely holds
  (names elsewhere, a full def for KU's Ku<->Ku ORGANIZES).
- base_analytics_service._to_domain_model: method-level TypeVar (was wrongly
  bound to the class T).
- Narrowing: calendar tags `list(...)`, principles options/conflicts `tuple(...)`,
  conversion extra_fields `dict[str, Any]`, system_service int() cast,
  habits_scheduling load dict `float`, events_learning isoformat + FilterParams,
  learning_recommendation session_plan `dict[str, Any]`.

Verified: `mypy --enable-error-code arg-type core` = 0; default `mypy core`
Success; `./dev quality` green; 441 targeted tests pass.

+++ app/core/services/habits/habits_scheduling_service.py
@@ -906,7 +906,7 @@ class HabitsSchedulingService(BaseService[HabitsOperations, Habit]):
 
     def _generate_load_recommendations(
         self,
-        load_by_day: dict[str, int],
+        load_by_day: dict[str, float],
         avg_load: float,
     ) -> list[str]:
         """Generate recommendations based on load distribution."""
